+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.1.11"></a>
+## [0.1.11](https://github.com/newsuk/times-components/compare/@times-components/icons@0.1.10...@times-components/icons@0.1.11) (2018-02-15)
+
+
+
+
+**Note:** Version bump only for package @times-components/icons
+
 <a name="0.1.10"></a>
 ## [0.1.10](https://github.com/newsuk/times-components/compare/@times-components/icons@0.1.9...@times-components/icons@0.1.10) (2018-02-15)