+++ packages/link/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.13.21"></a>
+## [0.13.21](https://github.com/newsuk/times-components/compare/@times-components/link@0.13.20...@times-components/link@0.13.21) (2018-02-15)
+
+
+
+
+**Note:** Version bump only for package @times-components/link
+
 <a name="0.13.20"></a>
 ## [0.13.20](https://github.com/newsuk/times-components/compare/@times-components/link@0.13.19...@times-components/link@0.13.20) (2018-02-15)